+++ rollup.config.js
@@ -7,27 +7,29 @@ export default {
     output: {
         name: "lib",
         file: "dist/lips.js",
-        format: "iife"
+        format: "iife",
+        globals: {
+            "regeneratorRuntime": "_regenerator"
+        }
     },
     plugins: [
-        commonjs({
-            include: "node_modules/**"
-        }),
-        nodeResolve({
-            jsnext: true,
-            main: false
-        }),
         babel({
             "babelrc": false,
             "runtimeHelpers": true,
             "plugins": [
                 "@babel/plugin-transform-async-to-generator",
-                ["@babel/plugin-transform-runtime", {useESModules: true}]
+                "@babel/plugin-transform-regenerator"
             ],
             "presets": [
                 "@babel/preset-env"
             ],
             "exclude": "node_modules/**"
+        }),
+        nodeResolve({
+            mainFields: ["jsnext:main"]
+        }),
+        commonjs({
+            include: "node_modules/**"
         })
     ]
 };